\name{classed}
\alias{classed}
\alias{tfclass}
\alias{tfclass<-}     

\title{Tframe Library Utilities}
\description{Some utilities used by other functions in the tframe library.}

\usage{
    classed(x, cls)
    tfclass(x)
    tfclass(x) <- str
}
\arguments{
    \item{x}{An object.}
    \item{cls}{A vector of strings.}
    \item{str}{A vector of strings.}
}
\keyword{internal}

