\name{print.tfPADIdata}
\alias{print.tfPADIdata}

\title{Print a tfPADIdata Object}
\description{See the generic function description.}

\usage{
	print(x)
}
\arguments {
	\item{x}{A tfPADIdata object.}
}
\sideffects{The object is printed.}
\seealso{
	\code{\link[base]{print}}
	\code{\link{tfPADIdata}}
}
\examples{
	print(tfPADIdata(c("series1", "series2"), server="ets"))
}
\keyword{tframe}
\keyword{ts}

