\name{tfputpadi}
\alias{tfputpadi}

\title{Write Data to a Data Server}
\description{Write data to a server.}
\usage{
	tfputpadi(x)
}
\value{A tfPADIdata object.}

\details{
	This function writes data to 
	a database using the TS PADI programs available 
	at <www.bank-banque-canada.ca/pgilbert> . 
}

\seealso{ 
	\code{\link{tfPADIdata}}
	\code{\link[padi]{putpadi}}
}
\keyword{tframe}
\keyword{time series data}

