\name{tsmatrix}
\alias{tsmatrix}
\alias{tsmatrix.default}
\title{Create a time series matrix.}
\description{
    This function is defunct. Consider using tbind.}
\usage{
%    tsmatrix(x, ...)
%    \method{tsmatrix}{default}(x, ...)
}
\arguments{
    \item{x}{a matrix.}
    \item{...}{arguments passed to other methods.}
}
\details{
    The default for the generic is set to the original function. See the S/R
    documentation for the original.
}
\seealso{
    \code{\link{tbind}}
}
\keyword{internal}


