\name{classed}
\alias{classed}
\alias{tfclass}
\alias{tfclass<-}     
\alias{seqN}     

\title{Tframe Library Utilities}
\description{Some utilities used by other functions in the tframe library.}

\usage{
	classed(x, str)
	tfclass(x)
	tfclass(x) <- str
	seqN(N)
}

\arguments{
	\item{x}{An object.}
	\item{str}{A vector of strings.}
}
\keyword{tframe}
\keyword{ts}

