\name{lag}
\alias{lag}
\alias{lag.default}
\title{Lag}
\description{lag a tframed object.}
\usage{
    lag(x, ...)
    \method{lag}{default}(x, ...)
}
\arguments{
    \item{x}{a tframed object.}
    \item{...}{arguments passed to other methods.}
}
\details{
    See documentation for the generic (from library ts in R).
}
\seealso{
    \code{\link{lag}}
    \code{\link{lag.default}}
}
\keyword{programming}
\keyword{utilities}
\keyword{ts}
\keyword{chron}


