\name{lag}
\alias{lag}
\alias{lag.default}
\title{Lag}
\description{lag a tframed object.}
\usage{
%    lag(x, ...)
%    \method{lag}{default}(x, ...)
}
\arguments{
    \item{x}{a tframed object.}
    \item{...}{arguments passed to other methods.}
}
\details{
    See documentation for the generic (from library ts in R).
}
\seealso{
    \code{\link{lag}}
    \code{\link{lag.default}}
}
\keyword{internal}


%\name{tsmatrix}
%\alias{tsmatrix}
%\alias{tsmatrix.default}
%\title{Create a time series matrix.}
%\description{
%    This function is defunct. Consider using tbind.}
%\usage{
%    tsmatrix(x, ...)
%    \method{tsmatrix}{default}(x, ...)
%}
%\arguments{
%    \item{x}{a matrix.}
%    \item{...}{arguments passed to other methods.}
%}
%\details{
%    The default for the generic is set to the original function. See the S/R
%    documentation for the original.
%}
%\seealso{
%    \code{\link{tbind}}
%}
%\keyword{internal}


