\name{lag}
\alias{lag}
\alias{lag.tframe}
\alias{tsmatrix}
\alias{tsmatrix.default}

\title{Misc Generic Replacements}
\description{
	These functions replace functions in S or R with generics if necessary.}
\usage{
	lag(x)
	tsmatrix(,x,y,z)
}

\details{
	The default for the generic is set to the original function. See the S/R
	documentation for the original.
}

\keyword{tframe}
\keyword{ts}
